+++ python/python/glide/glide_client.py
@@ -295,9 +295,9 @@ class BaseClient(CoreCommands):
                 command.args_array.args[:] = encoded_args
             else:
                 command.args_vec_pointer = create_leaked_bytes_vec(encoded_args)
-            batch_commands.append(command)
-        request.batch.commands.extend(batch_commands)
-        request.batch.is_atomic = is_atomic
+            transaction_commands.append(command)
+        request.batch.commands.extend(transaction_commands)
+        request.batch.is_atomic = True
         set_protobuf_route(request, route)
         return await self._write_request_await_response(request)